Implement Authentication Middleware #345
